dom/media/webaudio/test/test_channelMergerNodeWithVolume.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test ChannelMergerNode</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="webaudio.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

var gTest = {
  length: 2048,
  numberOfChannels: 6,
  createGraph: function(context) {
    var buffers = [];
    for (var j = 0; j < 6; ++j) {
      var buffer = context.createBuffer(2, 2048, context.sampleRate);
      for (var i = 0; i < 2048; ++i) {
        buffer.getChannelData(0)[i] = Math.sin(440 * 2 * (j + 1) * Math.PI * i / context.sampleRate);
        // Second channel is silent
      }
      buffers.push(buffer);
    }

    var merger = context.createChannelMerger();
    is(merger.channelCount, 1, "merger node has 1 input channels");
    is(merger.channelCountMode, "explicit", "Correct channelCountMode for the merger node");
    is(merger.channelInterpretation, "speakers", "Correct channelCountInterpretation for the merger node");

    for (var i = 0; i < 6; ++i) {
      var source = context.createBufferSource();
      source.buffer = buffers[i];
      var gain = context.createGain();
      gain.gain.value = 0.5;
      source.connect(gain);
      gain.connect(merger, 0, i);
      source.start(0);
    }

    return merger;
  },
  createExpectedBuffers: function(context) {
    var expectedBuffer = context.createBuffer(6, 2048, context.sampleRate);
    for (var i = 0; i < 6; ++i) {
      for (var j = 0; j < 2048; ++j) {
        expectedBuffer.getChannelData(i)[j] = 0.5 * 0.5 * Math.sin(440 * 2 * (i + 1) * Math.PI * j / context.sampleRate);
      }
    }
    return expectedBuffer;
  },
};

runTest();

</script>
</pre>
</body>
</html>
